migrate auth/forgot-password/*, auth/login, auth/logout
